Widen mmap and pread/pwrite offset and length to off_t / size_t
sys/mman.h declared mmap / munmap / mremap / msync / mprotect with an
`int` length and mmap with an `int` file offset. POSIX types these as
size_t / off_t, both 64-bit; an `int` declaration truncates a mapping
at or past 2GB and a file offset past 2GB. unistd.h's plain pread /
pwrite had the same `int` offset (and an `int` byte count), unlike the
pread64 / pwrite64 variants that already use the wide types.

Declare the lengths as `unsigned long` and the offsets as `long`,
matching size_t / off_t on the LP64 targets these POSIX blocks serve,
so a large-file memory map or positioned read/write is not truncated.

No deterministic regression is possible without a multi-gigabyte file;
the parameter widening uses the same path covered by the time/file
width fixtures, and the demos exercise mmap / pread at normal sizes.

+++ headers/include/unistd.h
@@ -184,10 +184,13 @@ extern char **environ;
 
 int open(char *path, int flags, ...);
 int read(int fd, char *buf, int n);
-int pread(int fd, char *buf, int n, int offset);
+// POSIX: pread/pwrite take an off_t offset and size_t count; an `int`
+// offset truncates positions past 2GB. Matches the pread64/pwrite64
+// signatures below.
+long pread(int fd, char *buf, unsigned long n, long offset);
 int close(int fd);
 int write(int fd, char *buf, int n);
-int pwrite(int fd, char *buf, int n, int offset);
+long pwrite(int fd, char *buf, unsigned long n, long offset);
 #ifdef __linux__
 // glibc large-file variants (`_LARGEFILE64_SOURCE`). The offset and
 // result are 64-bit; programs configured with `USE_PREAD64` (e.g.
